+++ engine/combat/appraisals/valere/moonerang.py
@@ -23,7 +23,7 @@ class Moonerang(SoSAppraisal):
 
     def __init__(
         self: Self,
-        value: int = 0,
+        value: float = 0.0,
         timing_type: SoSTimingType = SoSTimingType.MultiHit,
         boost: int = 0,
     ) -> None:
