+++ engine/combat/appraisals/zale/sunball.py
@@ -19,7 +19,7 @@ logger = logging.getLogger(__name__)
 class Sunball(SoSAppraisal):
     def __init__(
         self: Self,
-        value: int = 0,
+        value: float = 0.0,
         hold_time: float = 4.0,
         timing_type: SoSTimingType = SoSTimingType.Charge,
         boost: int = 0,
